libiff: clear neutrino strict-warning errors in reader + chunk iterators
Standalone build under neutrino's strict warning set
(-Wuseless-cast / -Wconversion / -Wunused-parameter / -Werror) was
failing on these:

  input.cc      reader::tell() / reader::size() each used
                const_cast<std::istream&>(m_stream) before calling
                tellg() / seekg(). The cast is genuinely useless —
                m_stream is a non-const reference and the enclosing
                method's const-ness does not propagate through a
                reference member, so the underlying istream is
                already mutable from those call sites.

  chunk_reader.cc  read_bytes() wrapped remaining() in
                static_cast<std::size_t>(...). remaining() already
                returns std::size_t.

  iff85_chunk_iterator.cc  When clamping a uint32_t chunk_size to
                m_options.max_chunk_size (uint64_t), the implicit
                truncation triggered -Werror=conversion. Made the
                narrowing explicit with static_cast<std::uint32_t>.

  riff_chunk_iterator.cc  get_size_override()'s 'offset' parameter
                is currently unreferenced. Marked with /*offset*/ so
                the param-name is still self-documenting at the
                declaration but doesn't trip -Werror=unused-parameter.

No behavioural change; build is clean under the strict warning set.

+++ src/libiff/chunk_reader.cc
@@ -51,7 +51,7 @@ namespace iff {
     }
     
     std::vector<std::byte> chunk_reader::read_bytes(std::size_t n) {
-        std::size_t to_read = std::min(n, static_cast<std::size_t>(remaining()));
+        std::size_t to_read = std::min(n, remaining());
         std::vector<std::byte> result(to_read);
         
         if (to_read > 0) {
